Section 19, Lecture 258 - Coding Challenge (WIP)
This commit is a WIP. In the middle of setting up all of the
TableView DataSource methods, and TableView Data manipulation
methods for the Category tableview. Added a new entity to
the CoreData datamodel document (Category) and setup the one
to many relationship between a category and its assoc. items.